next/store: some refactoring and fixing an off-by-one corner case
