+++ src/packages/next/components/misc/timestamp.tsx
@@ -1,3 +1,8 @@
+/*
+ *  This file is part of CoCalc: Copyright © 2022 Sagemath, Inc.
+ *  License: AGPLv3 s.t. "Commons Clause" – see LICENSE.md for details
+ */
+
 import { Tooltip } from "antd";
 import { CSSProperties } from "react";
 import TimeAgo from "timeago-react";
@@ -10,41 +15,48 @@ interface Props {
   absolute?: boolean;
 }
 
-export default function Timestamp({
-  epoch,
-  datetime,
-  style,
-  absolute,
-  dateOnly,
-}: Props) {
+export default function Timestamp(props: Props) {
+  const { epoch, style, absolute, dateOnly } = props;
+  let datetime = props.datetime;
+
   if (epoch && datetime == null) {
     datetime = new Date(epoch);
   }
+
   if (!datetime) {
     return <span style={style}>-</span>;
   }
+
   if (typeof datetime == "string") {
     datetime = new Date(datetime);
     if (typeof datetime == "string") throw Error("bug");
   }
-  const absoluteTime = dateOnly
-    ? datetime.toLocaleDateString(undefined, {
-        year: "numeric",
-        month: "short",
-        day: "numeric",
-      })
-    : datetime.toLocaleString(undefined, {
-        year: "numeric",
-        month: "short",
-        day: "numeric",
-        hour: "numeric",
-        minute: "numeric",
-      });
+
+  const absoluteTimeDateOnly = datetime.toLocaleDateString(undefined, {
+    year: "numeric",
+    month: "short",
+    day: "numeric",
+  });
+  const absoluteTimeFull = datetime.toLocaleString(undefined, {
+    year: "numeric",
+    month: "short",
+    day: "numeric",
+    hour: "numeric",
+    minute: "numeric",
+  });
+
+  const timeShown = dateOnly ? absoluteTimeDateOnly : absoluteTimeFull;
+
   if (absolute) {
-    return <span style={style}>{absoluteTime}</span>;
+    return (
+      <Tooltip trigger={["hover", "click"]} title={absoluteTimeFull}>
+        <span style={style}>{timeShown}</span>
+      </Tooltip>
+    );
   }
+
   return (
-    <Tooltip trigger={["hover", "click"]} title={absoluteTime}>
+    <Tooltip trigger={["hover", "click"]} title={timeShown}>
       <TimeAgo style={style} datetime={datetime} />
     </Tooltip>
   );
