+++ src/packages/next/components/store/cart.tsx
@@ -25,14 +25,10 @@ import useAPI from "lib/hooks/api";
 import useIsMounted from "lib/hooks/mounted";
 import { useRouter } from "next/router";
 import { useEffect, useMemo, useState } from "react";
+import { computeCost } from "./compute-cost";
 import OtherItems from "./other-items";
 import { EditRunLimit } from "./run-limit";
-import {
-  computeCost,
-  describeItem,
-  describePeriod,
-  DisplayCost,
-} from "./site-license-cost";
+import { describeItem, describePeriod, DisplayCost } from "./site-license-cost";
 
 export default function ShoppingCart() {
   const isMounted = useIsMounted();
