+++ src/packages/next/components/store/site-license-cost.tsx
@@ -4,13 +4,8 @@
  */
 
 import { Icon } from "@cocalc/frontend/components/icon";
-import { AVG_MONTH_DAYS } from "@cocalc/util/consts/billing";
 import { untangleUptime } from "@cocalc/util/consts/site-license";
 import { describe_quota } from "@cocalc/util/licenses/describe-quota";
-import {
-  compute_cost,
-  compute_cost_dedicated,
-} from "@cocalc/util/licenses/purchase/compute-cost";
 import {
   CostInputPeriod,
   PurchaseInfo,
@@ -19,135 +14,13 @@ import {
 import { money, percent_discount } from "@cocalc/util/licenses/purchase/utils";
 import { plural } from "@cocalc/util/misc";
 import { getDays } from "@cocalc/util/stripe/timecalcs";
-import { PRICES } from "@cocalc/util/upgrades/dedicated";
-import { ComputeCostProps } from "@cocalc/util/upgrades/shopping";
 import {
   dedicatedDiskDisplay,
   dedicatedVmDisplay,
 } from "@cocalc/util/upgrades/utils";
 import Timestamp from "components/misc/timestamp";
 import { ReactNode } from "react";
-
-function computeDedicatedDiskCost(
-  props: ComputeCostProps
-): CostInputPeriod | undefined {
-  if (props.type !== "disk") {
-    throw new Error("compute cost for disk only");
-  }
-  if (props.dedicated_disk == null)
-    throw new Error("missing props.dedicated_disk");
-  const { dedicated_disk } = props;
-  if (props.period != "monthly") throw new Error("period must be monthly");
-  if (dedicated_disk === false) throw new Error(`should not happen`);
-
-  try {
-    return {
-      input: { ...props, subscription: props.period },
-      ...compute_cost_dedicated({
-        dedicated_disk,
-        subscription: props.period,
-      }),
-    };
-  } catch (err) {
-    console.log(`problem calculating dedicated price: ${err}`);
-  }
-}
-
-function computeDedicatedVMCost(
-  props: ComputeCostProps
-): CostInputPeriod | undefined {
-  if (props.type !== "vm") {
-    throw new Error("compute cost for VM only");
-  }
-  if (props.dedicated_vm == null) {
-    throw new Error("missing props.dedicated_vm");
-  }
-  const { range, dedicated_vm } = props;
-  const machine = dedicated_vm.machine;
-  if (range == null || range[0] == null || range[1] == null) return;
-  const price_day = PRICES.vms[machine]?.price_day;
-  if (price_day == null) return;
-  const days = getDays({ start: range[0], end: range[1] });
-  const price = days * price_day;
-  return {
-    cost: price,
-    cost_per_unit: price,
-    discounted_cost: price,
-    cost_per_project_per_month: AVG_MONTH_DAYS * price_day,
-    cost_sub_month: AVG_MONTH_DAYS * price_day,
-    cost_sub_year: 12 * AVG_MONTH_DAYS * price_day,
-    input: {
-      ...props,
-      subscription: "no",
-      start: range[0] ?? new Date(),
-      end: range?.[1],
-    },
-    period: "range",
-  };
-}
-
-export function computeCost(
-  props: ComputeCostProps
-): CostInputPeriod | undefined {
-  const type = props.type ?? "quota";
-  switch (type) {
-    case "disk":
-      return computeDedicatedDiskCost(props);
-
-    case "vm":
-      return computeDedicatedVMCost(props);
-
-    case "quota":
-    default:
-      if (props.type === "disk" || props.type === "vm") {
-        throw Error("must be a quota upgrade license");
-      }
-      const {
-        user,
-        run_limit,
-        period,
-        range,
-        ram,
-        cpu,
-        disk,
-        always_running,
-        member,
-        uptime,
-        boost = false, // if true, allow "all zero" values and start at 0 USD
-      } = props;
-
-      if (period == "range" && range?.[1] == null) {
-        return undefined;
-      }
-
-      const input: PurchaseInfo = {
-        type: "quota",
-        user,
-        upgrade: "custom" as "custom",
-        quantity: run_limit,
-        subscription: (period == "range" ? "no" : period) as
-          | "no"
-          | "monthly"
-          | "yearly",
-        start: range?.[0] ?? new Date(),
-        end: range?.[1],
-        custom_ram: ram,
-        custom_dedicated_ram: 0,
-        custom_cpu: cpu,
-        custom_dedicated_cpu: 0,
-        custom_disk: disk,
-        custom_always_running: always_running,
-        custom_member: member,
-        custom_uptime: uptime,
-        boost,
-      };
-      return {
-        ...compute_cost(input),
-        input,
-        period,
-      };
-  }
-}
+import { useTimeFixer } from "./util";
 
 interface Props {
   cost: CostInputPeriod;
@@ -155,7 +28,8 @@ interface Props {
   oneLine?: boolean;
 }
 
-export function DisplayCost({ cost, simple, oneLine }: Props) {
+export function DisplayCost(props: Props) {
+  const { cost, simple, oneLine } = props;
   if (isNaN(cost.cost) || isNaN(cost.discounted_cost)) {
     return <>&ndash;</>;
   }
@@ -233,7 +107,9 @@ export function describeItem(info: Partial<PurchaseInfo>): ReactNode {
   if (info.quantity == null) {
     throw new Error("should not happen");
   }
-  const { always_running, idle_timeout } = untangleUptime(info.custom_uptime ?? "short");
+  const { always_running, idle_timeout } = untangleUptime(
+    info.custom_uptime ?? "short"
+  );
   return (
     <>
       {describe_quota({
@@ -254,17 +130,26 @@ function describeQuantity({ quantity = 1 }: { quantity?: number }): ReactNode {
   return `for ${quantity} running ${plural(quantity, "project")}`;
 }
 
-export function describePeriod({
-  subscription,
-  start,
-  end,
-}: {
+interface PeriodProps {
   subscription?: Omit<Subscription, "no">;
   start?: Date | string;
   end?: Date | string;
-}): ReactNode {
+}
+
+export function describePeriod(props: PeriodProps): ReactNode {
+  const { subscription, start: startRaw, end: endRaw } = props;
+
+  const { fromServerTime } = useTimeFixer();
+
   if (subscription == "no") {
-    if (start == null || end == null) throw new Error(`start date not set!`);
+    if (startRaw == null || endRaw == null)
+      throw new Error(`start date not set!`);
+    // we do not use startOfDay and endOfDay, because this was already
+    // done in "usage-and-duration::fixRangeSelector"
+    // rather, we calculate back to the user's offset
+    const start = fromServerTime(startRaw);
+    const end = fromServerTime(endRaw);
+
     const days = getDays({ start, end });
     return (
       <>
