+++ src/packages/next/components/store/usage-and-duration.tsx
@@ -6,12 +6,12 @@
 import { COSTS } from "@cocalc/util/licenses/purchase/consts";
 import { Subscription } from "@cocalc/util/licenses/purchase/types";
 import { endOfDay, startOfDay } from "@cocalc/util/stripe/timecalcs";
+import { DateRangeType } from "@cocalc/util/types/store";
 import { Divider, Form, Input, Radio, Space } from "antd";
 import A from "components/misc/A";
 import DateRange from "components/misc/date-range";
 import { ReactNode } from "react";
 import { useTimeFixer } from "./util";
-import { DateRangeType } from "@cocalc/util/types/store";
 
 interface Props {
   showExplanations?: boolean;
