+++ src/packages/util/stripe/timecalcs.ts
@@ -1,8 +1,16 @@
+/*
+ *  This file is part of CoCalc: Copyright © 2022 Sagemath, Inc.
+ *  License: AGPLv3 s.t. "Commons Clause" – see LICENSE.md for details
+ */
+
 import moment from "moment";
 import { ONE_DAY_MS } from "@cocalc/util/consts/billing";
-import { StartEndDatesWithStrings } from "../licenses/purchase/types";
+import { StartEndDatesWithStrings } from "@cocalc/util/licenses/purchase/types";
 
-// this does NOT rounds start/end of the day. it assumes this is done by the frontend form (using the user's time zone!)
+// this does NOT rounds start/end of the day.
+// take special care if you do this in the front-end, because if server-time is off by a significant amount,
+// it might count one day too many or too little.
+// use the rounding functions below to fix this, but maybe you have to use the "store/util::useTimeFixer" hook.
 export function getDays({ start, end }: StartEndDatesWithStrings): number {
   if (start == null || end == null) {
     throw Error("bug");
